fix: sidocode type

See merge request s09-final/S09P31B209!301

+++ Backend/src/main/java/com/ssafy/flowerly/address/repository/SigunguRepository.java
@@ -17,7 +17,7 @@ public interface SigunguRepository extends JpaRepository<Sigungu, Long > {
 
     Optional<Sigungu> findBySigunguNameAndSido(String sigunguName, Sido sido);
 
-    Optional<Sigungu> findBySigunguCodeAndSido(Integer sigunguCode, Sido sido);
+    Optional<Sigungu> findBySigunguCodeAndSido(Long sigunguCode, Sido sido);
 
     @Query(
             "Select s FROM Sigungu s " +
